move other user cursor highlighting to new file
