+++ frontend/src/app/composition_tool/CompositionTool.tsx
@@ -22,6 +22,7 @@ import { DocumentMetadata, ShareStyle } from "../lib/src/documentProperties";
 import { Comment } from "../lib/src/Comment";
 import { OnlineEntity } from "../lib/src/realtimeUserTypes";
 import { SelectedNote } from "../lib/src/SelectedNote";
+import { processOnlineUsersCursors } from './onlineUsers/updateUserCursors';
 
 import * as d3 from 'd3';
 import { Selection } from 'd3';
@@ -1427,61 +1428,12 @@ export default function CompositionTool() {
     };
 
     useEffect(() => {
-        // First, clear previous highlighting for other users
-        d3.selectAll('.other-user-highlight').each(function () {
-            d3.select(this).style('fill', null);
-            d3.select(this).classed('other-user-highlight', false);
-        });
-
-        // Iterate over onlineUsers
-        onlineUsers.forEach((onlineEntity, user_id) => {
-            // Exclude the current user
-            if (user_id !== userId.current) {
-                const cursor = onlineEntity.cursor as SelectedNote;
-                if (cursor && cursor.noteID && cursor.color) {
-                    const noteHeadId = cursor.noteID;
-                    const color = cursor.color;
-
-                    // Select the notehead element by its CSS ID
-                    const noteHeadElement = d3.select(`[id="${noteHeadId}"]`);
-                    if (!noteHeadElement.empty()) {
-                        noteHeadElement
-                            .style('fill', color)
-                            .classed('other-user-highlight', true);
-                    } else {
-                        console.warn(`Notehead with ID ${noteHeadId} not found`);
-                    }
-                }
-            }
-        });
-
-        const updateUserList = async () => {
-            console.log('Running updateUserList!');
-            const users: { userId: string; displayName: string; color: string }[] = [];
-
-            const promises = [];
-
-            onlineUsers.forEach((onlineEntity, userIdKey) => {
-                console.log(`Is ${userIdKey} !== ${userId.current}?`);
-                if (userIdKey !== userId.current) {
-                    const cursor = onlineEntity.cursor as SelectedNote;
-                    if (cursor && cursor.color) {
-                        const color = cursor.color;
-                        const displayNamePromise = fetchDisplayName(userIdKey).then((displayName) => {
-                            users.push({ userId: userIdKey, displayName, color });
-                        });
-                        promises.push(displayNamePromise);
-                    }
-                }
-            });
-
-            // await Promise.all(promises);
-            // setUserList(users);
-            userList.current = users;
-        };
-
-
-        updateUserList();
+        processOnlineUsersCursors(
+            userId.current,
+            onlineUsers,
+            userList,
+            fetchDisplayName
+        )
     }, [onlineUsers]);
 
     return (
